requirements.txt
py-solc>=3.2
web3>=4.8
